requirements.txt
# Simple PDF Processing
PyMuPDF>=1.23.0
pdf2image>=1.16.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
fitz>=0.16.0
# Image Processing (Optional)
opencv-python>=4.8.0
Pillow>=9.5.0

# Text Processing
unicodedata2>=15.0.0

# Utilities
tqdm>=4.64.0
pandas>=1.5.0
pathlib2>=2.3.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0